requirements.txt
flask==2.3.3
streamlit==1.28.0
pdf2image==1.16.3
Pillow==10.0.1
requests==2.31.0
pandas==2.1.1
openpyxl==3.1.2
python-docx==0.8.11
PyPDF2==3.0.1
watchdog==3.0.0